Renamed ERC721Basic to ERC721, and ERC721 to ERC721Full.
