+++ contracts/mocks/ERC721MintableBurnableImpl.sol
@@ -1,6 +1,6 @@
 pragma solidity ^0.4.24;
 
-import "../token/ERC721/ERC721.sol";
+import "../token/ERC721/ERC721Full.sol";
 import "../token/ERC721/ERC721Mintable.sol";
 import "../token/ERC721/ERC721Burnable.sol";
 
@@ -8,10 +8,10 @@ import "../token/ERC721/ERC721Burnable.sol";
 /**
  * @title ERC721MintableBurnableImpl
  */
-contract ERC721MintableBurnableImpl is ERC721, ERC721Mintable, ERC721Burnable {
+contract ERC721MintableBurnableImpl is ERC721Full, ERC721Mintable, ERC721Burnable {
   constructor()
     ERC721Mintable()
-    ERC721("Test", "TEST")
+    ERC721Full("Test", "TEST")
     public
   {
   }
