+++ contracts/token/ERC721/ERC721Metadata.sol
@@ -1,10 +1,10 @@
 pragma solidity ^0.4.24;
 
-import "./ERC721Basic.sol";
+import "./ERC721.sol";
 import "./IERC721Metadata.sol";
 import "../../introspection/ERC165.sol";
 
-contract ERC721Metadata is ERC165, ERC721Basic, IERC721Metadata {
+contract ERC721Metadata is ERC165, ERC721, IERC721Metadata {
   // Token name
   string internal name_;
