+++ contracts/token/ERC721/IERC721Enumerable.sol
@@ -1,13 +1,13 @@
 pragma solidity ^0.4.24;
 
-import "./IERC721Basic.sol";
+import "./IERC721.sol";
 
 
 /**
  * @title ERC-721 Non-Fungible Token Standard, optional enumeration extension
  * @dev See https://github.com/ethereum/EIPs/blob/master/EIPS/eip-721.md
  */
-contract IERC721Enumerable is IERC721Basic {
+contract IERC721Enumerable is IERC721 {
   function totalSupply() public view returns (uint256);
   function tokenOfOwnerByIndex(
     address _owner,
