+++ contracts/token/ERC721/IERC721Metadata.sol
@@ -1,13 +1,13 @@
 pragma solidity ^0.4.24;
 
-import "./IERC721Basic.sol";
+import "./IERC721.sol";
 
 
 /**
  * @title ERC-721 Non-Fungible Token Standard, optional metadata extension
  * @dev See https://github.com/ethereum/EIPs/blob/master/EIPS/eip-721.md
  */
-contract IERC721Metadata is IERC721Basic {
+contract IERC721Metadata is IERC721 {
   function name() external view returns (string _name);
   function symbol() external view returns (string _symbol);
   function tokenURI(uint256 _tokenId) public view returns (string);
